+++ tests/Test/FragmentSpec.hs
@@ -50,17 +50,17 @@ hasErrors :: Value -> Bool
 hasErrors (Object object') = HashMap.member "errors" object'
 hasErrors _ = True
 
-shirtType :: ObjectType IO
-shirtType = ObjectType "Shirt" Nothing
+shirtType :: Out.ObjectType IO
+shirtType = Out.ObjectType "Shirt" Nothing
     $ HashMap.singleton resolverName
-    $ Field Nothing (ScalarOutputType string) mempty resolve
+    $ Out.Field Nothing (Out.NamedScalarType string) mempty resolve
   where
     (Schema.Resolver resolverName resolve) = size
 
-hatType :: ObjectType IO
-hatType = ObjectType "Hat" Nothing
+hatType :: Out.ObjectType IO
+hatType = Out.ObjectType "Hat" Nothing
     $ HashMap.singleton resolverName
-    $ Field Nothing (ScalarOutputType int) mempty resolve
+    $ Out.Field Nothing (Out.NamedScalarType int) mempty resolve
   where
     (Schema.Resolver resolverName resolve) = circumference
 
@@ -69,9 +69,9 @@ toSchema (Schema.Resolver resolverName resolve) = Schema
     { query = queryType, mutation = Nothing }
   where
     unionMember = if resolverName == "Hat" then hatType else shirtType
-    queryType = ObjectType "Query" Nothing
+    queryType = Out.ObjectType "Query" Nothing
         $ HashMap.singleton resolverName
-        $ Field Nothing (ObjectOutputType unionMember) mempty resolve
+        $ Out.Field Nothing (Out.NamedObjectType unionMember) mempty resolve
 
 spec :: Spec
 spec = do